pyproject.toml
[project]
name = "scraper-dados-financeiro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gspread>=6.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "selenium>=4.39.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dados_financeiros"]

[project.scripts]
main = "dados_financeiros.main:main"

[tool.ruff]
line-length = 120
select = ["E", "F", "W", "C90", "B", "ANN", "COM", "N", "Q", "S", "UP", "I"]
ignore = ["E501", "N812", "COM812"]
exclude = ["build/", "dist/", ".venv/"]
